+++ tests/tests_resources/data_for_testing.py
@@ -7,6 +7,7 @@ from flexiznam.config import PARAMETERS
 MOUSE_ID = "6437dcb13ded9c65df142a12"
 TEST_PROJECT = "demo_project"
 PROJECT_ID = "610989f9a651ff0b6237e0f6"
+SESSION = "mouse_physio_2p_S20211102"
 DATA_ROOT = Path(PARAMETERS["data_root"]["raw"]) / TEST_PROJECT
 PROCESSED_ROOT = Path(PARAMETERS["data_root"]["processed"]) / TEST_PROJECT